added(new model ProductCart), fixed (association issues)

+++ server/db/models/product.js
@@ -22,7 +22,7 @@ const Product = db.define('product', {
     }
   },
   price: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.INTEGER,
     allowNull: false,
     notEmpty: true,
     validate: {
